To load two task at same time.
In this code, the images are fetched separately on two different tasks.
The code within the Task is asyncrhonous with await keyword so within that block the action flow of the code is suspeneded.
But the two tasks are in the syncrhonous code so both get run at the same time without the delay.
As a result we get the images displayed pretty quickly and faster.