Cancellable Task - Problem Statement
In this code, the view is getting loaded even if the user click back and come back again and no longer referring to the page. The task still loads in the background and executed.
This will slow down the application if there are such multiple task we need to keep on running.
Ideally we should cancel the task as long as the user is no longer on that page or view.